requirements/base.txt
asgiref==3.11.1
Django==6.0.4
django-allauth==65.16.0
django-environ==0.13.0
pillow==12.2.0
psycopg==3.3.3
sqlparse==0.5.5
